+++ index.js
@@ -1,4 +1,4 @@
-import SunnyDom from './src/SunnyDom.js';
+import SunnyDom from './src/SunnyDom/index.js';
 import App from './src/App.js';
 
 window.onload = () => {
