+++ src/Sunny/index.js
@@ -1,9 +1,16 @@
+/*
+ ** ony using SunnyElement
+ */
+
 import createElement from './createElement.js';
 import parseDomElement from './parseDomElement.js';
+import findElement from './findElement.js';
 
 const Sunny = {
   createElement,
   parseDomElement,
+  findElement,
 };
 
 export default Sunny;
+export { createElement, parseDomElement, findElement };
