--- src/SunnyDom.js
@@ -1,7 +1,0 @@
-import render from './render.js';
-
-const SunnyDom = {
-  render,
-};
-
-export default SunnyDom;
